Remove unnecesary logic
This results in a nearly 20% speedup for the get_bits_from_image function. Cython is now only 40% faster than Python, and FFmpeg is now the major bottleneck in code. I have reason to believe this function is close to maximal optimization, unless it is rewritten to scan multiple pixels at a time (which might not work because bits need to be in sequential order) or use a different method to load pixels.